+++ tests/scripts/ahi.tc.IR-BD.imagery_clean.sh
@@ -18,20 +18,20 @@
 
 #!/bin/bash
 
+# 160 files - 10 per channel
 run_procflow \
-    $GEOIPS_BASEDIR/test_data/test_data_sar/data/STAR_SAR_20181025203206_WP312018_31W_FIX_3km.nc \
+    $GEOIPS_BASEDIR/test_data/test_data_ahi_day/data/20200405_0000/HS_H08_20200405_0000_B*_FLDK_R*_S*.DAT \
     --procflow single_source \
-    --reader_name sar_winds_netcdf \
-    --product_name nrcs \
-    --compare_path "$GEOIPS_BASEDIR/geoips_packages/recenter_tc/tests/outputs/sar_<product>" \
+    --reader_name ahi_hsd \
+    --product_name IR-BD \
+    --compare_path "$GEOIPS_BASEDIR/geoips_packages/recenter_tc/tests/outputs/ahi.tc.IR-BD.imagery_clean" \
     --output_format imagery_clean \
     --filename_format tc_clean_fname \
     --adjust_area_def recenter_tc \
     --metadata_filename_format metadata_default_fname \
     --metadata_output_format metadata_default \
     --trackfile_parser bdeck_parser \
-    --trackfiles $GEOIPS/tests/sectors/tc_bdecks/bwp312018.dat
-
+    --trackfiles $GEOIPS/tests/sectors/tc_bdecks/bsh252020.dat
 ss_retval=$?
 
 exit $((ss_retval))