+++ tests/scripts/amsr2.tc.color37.imagery_clean.sh
@@ -19,17 +19,17 @@
 #!/bin/bash
 
 run_procflow \
-    $GEOIPS_BASEDIR/test_data/test_data_gpm/data/3B-HHR-L.MS.MRG.3IMERG.20200917-S170000-E172959.1020.V06B.RT-H5 \
+    $GEOIPS_BASEDIR/test_data/test_data_amsr2/data//*mbt* \
     --procflow single_source \
-    --reader_name  imerg_hdf5 \
-    --product_name Rain \
-    --compare_path "$GEOIPS_BASEDIR/geoips_packages/recenter_tc/tests/outputs/imerg_<product>" \
+    --reader_name  amsr2_netcdf \
+    --product_name color37 \
+    --compare_path "$GEOIPS_BASEDIR/geoips_packages/recenter_tc/tests/outputs/amsr2.tc.color37.imagery_clean" \
     --output_format imagery_clean \
     --filename_format tc_clean_fname \
     --metadata_filename_format metadata_default_fname \
     --metadata_output_format metadata_default \
     --trackfile_parser bdeck_parser \
-    --trackfiles $GEOIPS/tests/sectors/tc_bdecks/bal202020.dat \
+    --trackfiles $GEOIPS/tests/sectors/tc_bdecks/bio012020.dat \
     --adjust_area_def recenter_tc
 ss_retval=$?