+++ tests/scripts/gmi.tc.89pct.imagery_clean.sh
@@ -19,17 +19,19 @@
 #!/bin/bash
 
 run_procflow \
-    $GEOIPS_BASEDIR/test_data/test_data_saphir/data/MT1SAPSL1A__1.09_000_1_19_I_2021_02_09_00_30_03_2021_02_09_01_11_16_48144_48144_497_33_33_KUX_00.h5 \
+    $GEOIPS_BASEDIR/test_data/test_data_gpm/data/1B.GPM.GMI.TB2016.20200917-S171519-E172017.V05A.RT-H5 \
+    $GEOIPS_BASEDIR/test_data/test_data_gpm/data/1B.GPM.GMI.TB2016.20200917-S172019-E172517.V05A.RT-H5 \
+    $GEOIPS_BASEDIR/test_data/test_data_gpm/data/1B.GPM.GMI.TB2016.20200917-S172519-E173017.V05A.RT-H5 \
     --procflow single_source \
-    --reader_name  saphir_hdf5 \
-    --product_name "183-3H" \
-    --compare_path "$GEOIPS_BASEDIR/geoips_packages/recenter_tc/tests/outputs/saphir_<product>" \
+    --reader_name  gmi_hdf5 \
+    --product_name 89pct \
+    --compare_path "$GEOIPS_BASEDIR/geoips_packages/recenter_tc/tests/outputs/gmi.tc.89pct.imagery_clean" \
     --output_format imagery_clean \
     --filename_format tc_clean_fname \
     --metadata_filename_format metadata_default_fname \
     --metadata_output_format metadata_default \
     --trackfile_parser bdeck_parser \
-    --trackfiles $GEOIPS/tests/sectors/tc_bdecks/bsh192021.dat \
+    --trackfiles $GEOIPS/tests/sectors/tc_bdecks/bal202020.dat \
     --adjust_area_def recenter_tc
 ss_retval=$?