+++ tests/scripts/hy2b.tc.windspeed.imagery_clean.sh
@@ -19,18 +19,17 @@
 #!/bin/bash
 
 run_procflow \
-    $GEOIPS_BASEDIR/test_data/test_data_scat/data/metopc_byu_uhr/210421_0142_12730_SURIGAE_210421_0000.WRave3.nc \
+    ${GEOIPS_BASEDIR}/test_data/test_data_hy2/data/hscat_20211202_080644_hy_2b__15571_o_250_2204_ovw_l2.nc \
     --procflow single_source \
-    --reader_name  ascat_uhr_netcdf \
-    --product_name windbarbs \
-    --tc_template_yaml $GEOIPS/geoips/yaml_configs/sectors_dynamic/tc_web_ascatuhr_barbs_template.yaml \
-    --compare_path "$GEOIPS_BASEDIR/geoips_packages/recenter_tc/tests/outputs/ascat_uhr_<product>" \
+    --reader_name scat_knmi_winds_netcdf \
+    --product_name windspeed \
+    --compare_path "$GEOIPS_BASEDIR/geoips_packages/recenter_tc/tests/outputs/hy2b.tc.windspeed.imagery_clean" \
     --output_format imagery_clean \
     --filename_format tc_clean_fname \
     --metadata_filename_format metadata_default_fname \
     --metadata_output_format metadata_default \
     --trackfile_parser bdeck_parser \
-    --trackfiles $GEOIPS/tests/sectors/tc_bdecks/bwp022021.dat \
+    --trackfiles $GEOIPS/tests/sectors/tc_bdecks/bwp272021.dat \
     --adjust_area_def recenter_tc
 ss_retval=$?