+++ tests/scripts/imerg.tc.Rain.imagery_clean.sh
@@ -19,17 +19,17 @@
 #!/bin/bash
 
 run_procflow \
-    ${GEOIPS_BASEDIR}/test_data/test_data_smos/data/SM_OPER_MIR_SCNFSW_20200216T120839_20200216T135041_110_001_7.nc \
+    $GEOIPS_BASEDIR/test_data/test_data_gpm/data/3B-HHR-L.MS.MRG.3IMERG.20200917-S170000-E172959.1020.V06B.RT-H5 \
     --procflow single_source \
-    --reader_name smos_winds_netcdf \
-    --product_name windspeed \
-    --compare_path "$GEOIPS_BASEDIR/geoips_packages/recenter_tc/tests/outputs/smos_<product>" \
+    --reader_name  imerg_hdf5 \
+    --product_name Rain \
+    --compare_path "$GEOIPS_BASEDIR/geoips_packages/recenter_tc/tests/outputs/imerg.tc.Rain.imagery_clean" \
     --output_format imagery_clean \
     --filename_format tc_clean_fname \
     --metadata_filename_format metadata_default_fname \
     --metadata_output_format metadata_default \
     --trackfile_parser bdeck_parser \
-    --trackfiles $GEOIPS/tests/sectors/tc_bdecks/bsh162020.dat \
+    --trackfiles $GEOIPS/tests/sectors/tc_bdecks/bal202020.dat \
     --adjust_area_def recenter_tc
 ss_retval=$?