+++ tests/scripts/metopc_knmi_125.tc.windbarbs.imagery_clean.sh
@@ -19,17 +19,18 @@
 #!/bin/bash
 
 run_procflow \
-    ${GEOIPS_BASEDIR}/test_data/test_data_scat/data/oscat_250/oscat_20210209_022459_scasa1_23155_o_250_2202_ovw_l2.nc \
+    $GEOIPS_BASEDIR/test_data/test_data_scat/data/metopc_knmi_125/ascat_20210421_010000_metopc_12730_eps_o_coa_3203_ovw.l2.nc \
     --procflow single_source \
     --reader_name scat_knmi_winds_netcdf \
-    --product_name windspeed \
-    --compare_path "$GEOIPS_BASEDIR/geoips_packages/recenter_tc/tests/outputs/oscat_<product>" \
+    --product_name windbarbs \
+    --compare_path "$GEOIPS_BASEDIR/geoips_packages/recenter_tc/tests/outputs/metopc_knmi_125.tc.windbarbs.imagery_clean" \
+    --tc_template_yaml $GEOIPS/geoips/yaml_configs/sectors_dynamic/tc_web_ascat_high_barbs_template.yaml \
     --output_format imagery_clean \
     --filename_format tc_clean_fname \
     --metadata_filename_format metadata_default_fname \
     --metadata_output_format metadata_default \
     --trackfile_parser bdeck_parser \
-    --trackfiles $GEOIPS/tests/sectors/tc_bdecks/bsh192021.dat \
+    --trackfiles $GEOIPS/tests/sectors/tc_bdecks/bwp022021.dat \
     --adjust_area_def recenter_tc
 ss_retval=$?