+++ tests/scripts/sar.tc.nrcs.imagery_clean.sh
@@ -19,19 +19,19 @@
 #!/bin/bash
 
 run_procflow \
-    $GEOIPS_BASEDIR/test_data/test_data_scat/data/metopc_knmi_125/ascat_20210421_010000_metopc_12730_eps_o_coa_3203_ovw.l2.nc \
+    $GEOIPS_BASEDIR/test_data/test_data_sar/data/STAR_SAR_20181025203206_WP312018_31W_FIX_3km.nc \
     --procflow single_source \
-    --reader_name scat_knmi_winds_netcdf \
-    --product_name windbarbs \
-    --compare_path "$GEOIPS_BASEDIR/geoips_packages/recenter_tc/tests/outputs/metopc_knmi_125_<product>" \
-    --tc_template_yaml $GEOIPS/geoips/yaml_configs/sectors_dynamic/tc_web_ascat_high_barbs_template.yaml \
+    --reader_name sar_winds_netcdf \
+    --product_name nrcs \
+    --compare_path "$GEOIPS_BASEDIR/geoips_packages/recenter_tc/tests/outputs/sar.tc.nrcs.imagery_clean" \
     --output_format imagery_clean \
     --filename_format tc_clean_fname \
+    --adjust_area_def recenter_tc \
     --metadata_filename_format metadata_default_fname \
     --metadata_output_format metadata_default \
     --trackfile_parser bdeck_parser \
-    --trackfiles $GEOIPS/tests/sectors/tc_bdecks/bwp022021.dat \
-    --adjust_area_def recenter_tc
+    --trackfiles $GEOIPS/tests/sectors/tc_bdecks/bwp312018.dat
+
 ss_retval=$?
 
 exit $((ss_retval))