+++ tests/scripts/ssmis.tc.color89.imagery_clean.sh
@@ -19,19 +19,17 @@
 #!/bin/bash
 
 run_procflow \
-    $GEOIPS_BASEDIR/test_data/test_data_gpm/data/1B.GPM.GMI.TB2016.20200917-S171519-E172017.V05A.RT-H5 \
-    $GEOIPS_BASEDIR/test_data/test_data_gpm/data/1B.GPM.GMI.TB2016.20200917-S172019-E172517.V05A.RT-H5 \
-    $GEOIPS_BASEDIR/test_data/test_data_gpm/data/1B.GPM.GMI.TB2016.20200917-S172519-E173017.V05A.RT-H5 \
+    $GEOIPS_BASEDIR/test_data/test_data_ssmis/data/US058SORB-RAWspp.sdris_f16_d20200519_s084400_e102900_r85579_cfnoc.raw \
     --procflow single_source \
-    --reader_name  gmi_hdf5 \
-    --product_name 89pct \
-    --compare_path "$GEOIPS_BASEDIR/geoips_packages/recenter_tc/tests/outputs/gmi_<product>" \
+    --reader_name  ssmis_binary \
+    --product_name color89 \
+    --compare_path "$GEOIPS_BASEDIR/geoips_packages/recenter_tc/tests/outputs/ssmis.tc.color89.imagery_clean" \
     --output_format imagery_clean \
     --filename_format tc_clean_fname \
     --metadata_filename_format metadata_default_fname \
     --metadata_output_format metadata_default \
     --trackfile_parser bdeck_parser \
-    --trackfiles $GEOIPS/tests/sectors/tc_bdecks/bal202020.dat \
+    --trackfiles $GEOIPS/tests/sectors/tc_bdecks/bio012020.dat \
     --adjust_area_def recenter_tc
 ss_retval=$?