+++ tests/scripts/viirs.tc.Infrared-Gray.imagery_clean.sh
@@ -34,7 +34,7 @@ run_procflow \
     --procflow single_source \
     --reader_name viirs_netcdf \
     --product_name Infrared-Gray \
-    --compare_path "$GEOIPS_BASEDIR/geoips_packages/recenter_tc/tests/outputs/viirs_<product>" \
+    --compare_path "$GEOIPS_BASEDIR/geoips_packages/recenter_tc/tests/outputs/viirs.tc.Infrared-Gray.imagery_clean" \
     --output_format imagery_clean \
     --filename_format tc_clean_fname \
     --adjust_area_def recenter_tc \